pyproject.toml
[tool.poetry]
name = "have-you-ever"
version = "0.1.0"
description = "Have You ever game, customized using AI and user gets to ask questions once in a while."
authors = ["Pratik Singh <71725848+PratikSingh121@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.1.4"
langchain = "^0.1.16"
langchain-core = "^0.1.46"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.0"
flask-socketio = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
